Update GuestQueue.cs
Task 1A- enqueue function          
I have implemented the function in GuestQueue that allows for the new guest objects to be enqueued. This is done first by using tail as an index to place the new guest object into store. next tail is increased by one to indicate where the next enqueued value will go. And lastly, numitems is increased by one to reflect the current count.